The single core version is now using the channel extractor. It produces the same output, but is slower. Found an error in the code that I will remove with the next commit.
git-svn-id: svn+ssh://jop91.astron.nl/auto/home/jive_cvs/SOFTC/svn/trunk/sfxc@217 5736e3ea-d426-0410-8aaf-93e336ac3372

+++ include/Mark4_header.h
@@ -225,7 +225,7 @@ Mark4_header<T>::get_microtime(int track) {
   return microsecond(track) + 
          1000000*(second(track) + 
                   60 * (minute(track) + 
-                        60*hour(track)));
+                        60*(INT64)hour(track)));
 } 
 
 template <class T>
